fix: Resolve CI type-checking failures and improve type safety across codebase (#115)
- Replace bare type: ignore comments with specific error codes (e.g.,
  no-any-return, assignment, attr-defined) for precise suppression
- Add unused-ignore safety net to all type: ignore comments for
  cross-environment compatibility (local vs CI mypy configurations)
- Wrap untyped return values with np.asarray(), float(), int(), bool(),
  and typing.cast() to improve actual type safety
- Fix torch.stack type: ignore placement in adaptive_fusion.py
  (must be on opening line for mypy to associate correctly)
- Restore --ignore-missing-imports flag to CI mypy invocation
- Add torch model cache to CI for faster ML test runs
- Switch opencv-python to opencv-python-headless for CI compatibility

+++ src/omni_mercury_engine/cognitive/multi_agent_coordination.py
@@ -691,7 +691,7 @@ class ConsensusProtocol:
         if unanimous:
             final_decision = decisions[0]
             agreement_ratio = 1.0
-            confidence = np.mean([r.confidence for r in results])
+            confidence = float(np.mean([r.confidence for r in results]))
             dissenting = []
         else:
             # No consensus - default to anomaly for safety
